Add switching of version based on production api check; moved json_get and config file functions into new util.py file
